Bug 549697: Provide API to manage multiple preference stores
The pre-existing case had one preference store that the terminal
managed. In preparation for lots of new preferences, make the
terminal control take a preference store to get preferences from.

To be able to make this change, the settings available today need
to be all wrapped in a preference store. In particular that means
that the font's definition needs to be scoped in the preference
store too.

Change-Id: Ife04ffbc02efcefb7d779ab91f08b8f4d8e542c8