Bug 549697: Mark update font methods in Terminal as deprecated
There are two ways of specifying fonts in the terminal,
one via font definitions, and one via the font directly. The
method via the font directly was already deprecated in the public
API, this commit follows that deprecation chain through to
make it easier to identify which code paths are not likely
relevant anymore.

These methods could simply be removed as part of the major version
bump.

Change-Id: I05d746c429c4e6c3ced5ee7bb7a6172e0a255c2d